Enable IPv6 addresses to be resolved and specified as valid URI
An IPv6 address must be enclosed in [] when in a URI (i.e.
http://[:::::::]:80).  This change allows the '[' and ']' as characters
allowed as part of the authority so the underlying HTTP library can
receive a resolved IPv6 address and utilize.  The subsystem itself must
support utilizing, this change only allows it to be treated as valid.

(cherry picked from commit 3be5716723471579ff94704281f261716c8f0b8a)